Applied the new CMake scheme to Common
- share contains resources to be installed in share/O2
- changed directory structure
- use buckets
- reformat
