+++ CMakeLists.txt
@@ -183,7 +183,3 @@ WRITE_CONFIG_FILE(config.sh)
 configure_file(${CMAKE_SOURCE_DIR}/CTestCustom.cmake
     ${CMAKE_BINARY_DIR}/CTestCustom.cmake
     )
-
-# Create an automatically generated config header file to pass file paths to the application
-configure_file(${PROJECT_SOURCE_DIR}/Common/Header/AliceO2Config.h.in
-    ${CMAKE_CURRENT_SOURCE_DIR}/Common/Header/AliceO2Config.h)
