+++ Common/CMakeLists.txt
@@ -1,2 +1,3 @@
-add_subdirectory (Field)
+add_subdirectory(share)
 add_subdirectory(MathUtils)
+add_subdirectory(Field)
