+++ Common/Field/src/FieldLinkDef.h
@@ -1,5 +1,5 @@
 
-#ifdef __CINT__
+#ifdef __CLING__
 #pragma link off all globals;
 #pragma link off all classes;
 #pragma link off all functions;
