--- Common/Header/AliceO2Config.h.in
@@ -1,2 +1,0 @@
-#define O2PROTO1_MAGF_DIR "@CMAKE_SOURCE_DIR@/Common/Resources/maps/mfchebKGI_sym.root"
-#define O2PROTO1_MAGF_CREATEFIELDMAP_DIR "@CMAKE_SOURCE_DIR@/Common/Resources/maps/mfchebKGI_sym.root"
