+++ Common/MathUtils/src/Chebyshev3DCalc.cxx
@@ -2,10 +2,11 @@
 /// \brief Implementation of the Cheb3DCalc class
 /// \author ruben.shahoyan@cern.ch 09/09/2006
 
-#include "Chebyshev3DCalc.h"
+#include "MathUtils/Chebyshev3DCalc.h"
 #include <TSystem.h>  // for TSystem, gSystem
 #include "TNamed.h"   // for TNamed
 #include "TString.h"  // for TString, TString::EStripType::kBoth
+
 using namespace AliceO2::MathUtils;
 
 ClassImp(Chebyshev3DCalc)
@@ -26,7 +27,7 @@ Chebyshev3DCalc::Chebyshev3DCalc()
 {
 }
 
-Chebyshev3DCalc::Chebyshev3DCalc(const Chebyshev3DCalc& src)
+Chebyshev3DCalc::Chebyshev3DCalc(const Chebyshev3DCalc &src)
   : TNamed(src),
     mNumberOfCoefficients(src.mNumberOfCoefficients),
     mNumberOfRows(src.mNumberOfRows),
@@ -79,7 +80,7 @@ Chebyshev3DCalc::Chebyshev3DCalc(const Chebyshev3DCalc& src)
   }
 }
 
-Chebyshev3DCalc::Chebyshev3DCalc(FILE* stream)
+Chebyshev3DCalc::Chebyshev3DCalc(FILE *stream)
   : mNumberOfCoefficients(0),
     mNumberOfRows(0),
     mNumberOfColumns(0),
@@ -96,7 +97,7 @@ Chebyshev3DCalc::Chebyshev3DCalc(FILE* stream)
   loadData(stream);
 }
 
-Chebyshev3DCalc& Chebyshev3DCalc::operator=(const Chebyshev3DCalc& rhs)
+Chebyshev3DCalc &Chebyshev3DCalc::operator=(const Chebyshev3DCalc &rhs)
 {
   if (this != &rhs) {
     Clear();
@@ -146,7 +147,7 @@ Chebyshev3DCalc& Chebyshev3DCalc::operator=(const Chebyshev3DCalc& rhs)
   return *this;
 }
 
-void Chebyshev3DCalc::Clear(const Option_t*)
+void Chebyshev3DCalc::Clear(const Option_t *)
 {
   if (mTemporaryCoefficients2D) {
     delete[] mTemporaryCoefficients2D;
@@ -178,10 +179,10 @@ void Chebyshev3DCalc::Clear(const Option_t*)
   }
 }
 
-void Chebyshev3DCalc::Print(const Option_t*) const
+void Chebyshev3DCalc::Print(const Option_t *) const
 {
-  printf("Chebyshev parameterization data %s for 3D->1 function, precision: %e\n", 
-	 GetName(),mPrecision);
+  printf("Chebyshev parameterization data %s for 3D->1 function, precision: %e\n",
+         GetName(), mPrecision);
   int nmax3d = 0;
   for (int i = mNumberOfElementsBound2D; i--;) {
     if (mCoefficientBound2D0[i] > nmax3d) {
@@ -191,7 +192,7 @@ void Chebyshev3DCalc::Print(const Option_t*) const
   printf("%d coefficients in %dx%dx%d matrix\n", mNumberOfCoefficients, mNumberOfRows, mNumberOfColumns, nmax3d);
 }
 
-Float_t Chebyshev3DCalc::evaluateDerivative(int dim, const Float_t* par) const
+Float_t Chebyshev3DCalc::evaluateDerivative(int dim, const Float_t *par) const
 {
   int ncfRC;
   for (int id0 = mNumberOfRows; id0--;) {
@@ -225,7 +226,7 @@ Float_t Chebyshev3DCalc::evaluateDerivative(int dim, const Float_t* par) const
                     : chebyshevEvaluation1D(par[0], mTemporaryCoefficients1D, mNumberOfRows);
 }
 
-Float_t Chebyshev3DCalc::evaluateDerivative2(int dim1, int dim2, const Float_t* par) const
+Float_t Chebyshev3DCalc::evaluateDerivative2(int dim1, int dim2, const Float_t *par) const
 {
   Bool_t same = dim1 == dim2;
   int ncfRC;
@@ -258,9 +259,9 @@ Float_t Chebyshev3DCalc::evaluateDerivative2(int dim1, int dim2, const Float_t* 
     }
   }
   return (dim1 == 0 || dim2 == 0)
-           ? (same ? chebyshevEvaluation1Derivative2(par[0], mTemporaryCoefficients1D, mNumberOfRows)
-                   : chebyshevEvaluation1Derivative(par[0], mTemporaryCoefficients1D, mNumberOfRows))
-           : chebyshevEvaluation1D(par[0], mTemporaryCoefficients1D, mNumberOfRows);
+         ? (same ? chebyshevEvaluation1Derivative2(par[0], mTemporaryCoefficients1D, mNumberOfRows)
+                 : chebyshevEvaluation1Derivative(par[0], mTemporaryCoefficients1D, mNumberOfRows))
+         : chebyshevEvaluation1D(par[0], mTemporaryCoefficients1D, mNumberOfRows);
 }
 
 #ifdef _INC_CREATION_Chebyshev3D_
@@ -301,7 +302,7 @@ void Chebyshev3DCalc::saveData(FILE* stream) const
 }
 #endif
 
-void Chebyshev3DCalc::loadData(FILE* stream)
+void Chebyshev3DCalc::loadData(FILE *stream)
 {
   if (!stream) {
     Error("LoadData", "No stream provided.\nStop");
@@ -359,7 +360,7 @@ void Chebyshev3DCalc::loadData(FILE* stream)
     mCoefficients[i] = buffs.Atof();
   }
   // read precision
-  readLine(buffs,stream);
+  readLine(buffs, stream);
   mPrecision = buffs.Atof();
 
   // check end_of_data record
@@ -370,7 +371,7 @@ void Chebyshev3DCalc::loadData(FILE* stream)
   }
 }
 
-void Chebyshev3DCalc::readLine(TString& str, FILE* stream)
+void Chebyshev3DCalc::readLine(TString &str, FILE *stream)
 {
   while (str.Gets(stream)) {
     str = str.Strip(TString::kBoth, ' ');
@@ -415,7 +416,7 @@ void Chebyshev3DCalc::initializeColumns(int nc)
     delete[] mTemporaryCoefficients2D;
     mTemporaryCoefficients2D = 0;
   }
-  if (mNumberOfColumns) mTemporaryCoefficients2D = new Float_t[mNumberOfColumns];
+  if (mNumberOfColumns) { mTemporaryCoefficients2D = new Float_t[mNumberOfColumns]; }
 }
 
 void Chebyshev3DCalc::initializeElementBound2D(int ne)
@@ -453,7 +454,7 @@ void Chebyshev3DCalc::initializeCoefficients(int nc)
   }
 }
 
-Float_t Chebyshev3DCalc::chebyshevEvaluation1Derivative(Float_t x, const Float_t* array, int ncf)
+Float_t Chebyshev3DCalc::chebyshevEvaluation1Derivative(Float_t x, const Float_t *array, int ncf)
 {
   if (--ncf < 1) {
     return 0;
@@ -479,7 +480,7 @@ Float_t Chebyshev3DCalc::chebyshevEvaluation1Derivative(Float_t x, const Float_t
   return b0 - x * b1 - dcf0 / 2;
 }
 
-Float_t Chebyshev3DCalc::chebyshevEvaluation1Derivative2(Float_t x, const Float_t* array, int ncf)
+Float_t Chebyshev3DCalc::chebyshevEvaluation1Derivative2(Float_t x, const Float_t *array, int ncf)
 {
   if (--ncf < 2) {
     return 0;
