+++ Common/share/CMakeLists.txt
@@ -1,0 +1,4 @@
+install(
+    DIRECTORY maps
+    DESTINATION share/o2
+)
